Added Particle Systems as Childs of Rocket Prefab
Then we used those particle systems added to prefab as Instance of Particle systems in the Serialized fields crashParticles and successParticles of type ParticleSystem. This ensures whenever particle system is triggered, it triggers on Top of Rocket. We do this by attaching the prefab-attached Particle system instance and not the actual particle system objects.